requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
alabaster==0.7.16
alembic==1.13.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
argparse-dataclass==2.0.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
biopython==1.83
bleach==6.1.0
blinker==1.8.2
blosc2==2.7.0
cachetools==5.3.3
certifi==2024.6.2
certipy==0.1.3
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
colormath==3.0.0
comm==0.2.2
conda-inject==1.3.2
ConfigArgParse==1.7
connection_pool==0.0.3
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
Cython==3.0.10
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==5.5
datrie==0.8.2
debugpy==1.8.1
decorator==5.1.1
deepTools==3.5.5
deeptoolsintervals==0.1.9
defusedxml==0.7.1
docker-pycreds==0.4.0
docutils==0.21.2
dpath==2.2.0
emcee==3.1.6
entrypoints==0.4
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.3
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.31.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
humanfriendly==10.0
humanize==4.9.0
hydra-core==1.3.2
idna==3.7
imageio==2.34.1
imagesize==1.4.1
immutables==0.20
importlib_metadata==7.2.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.25.0
ipython-genutils==0.2.0
ipywidgets==8.1.3
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.30
jax-cuda12-pjrt==0.4.30
jax-cuda12-plugin==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_fileid==0.9.2
jupyter_server_proxy==4.2.0
jupyter_server_terminals==0.5.3
jupyter_server_ydoc==0.8.0
jupyterhub==5.0.0
jupyterlab==4.2.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_slurm==3.0.1
jupyterlab_widgets==3.0.11
kaleido==0.2.1
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lightning==2.3.0
lightning-utilities==0.11.2
llvmlite==0.43.0
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
medmnist==3.0.1
meson==1.4.1
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiqc==1.22.3
namex==0.0.8
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nbgitpuller==1.2.1
ndindex==1.8
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nose==1.3.7
notebook==6.5.7
notebook_shim==0.2.4
numba==0.60.0
numexpr==2.10.1
numpy==1.25.2
numpydoc==1.7.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.18.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openseespy==3.6.0.2
openseespylinux==3.6.0.2
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==24.1
pamela==1.1.0
pandas==2.2.2
pandoc==2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
plac==1.4.3
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
plumbum==1.8.3
ply==3.11
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
PuLP==2.8.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py2bit==0.3.0
pyaml-env==1.2.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyBigWig==0.3.23
PyBIS==1.36.3
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pyg-lib==0.4.0+pt22cu121
Pygments==2.18.0
Pyomo==6.7.3
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyquaternion==0.9.9
pysam==0.22.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-lightning==2.3.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
QtPy==2.4.1
rdflib==7.0.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
reretry==0.11.8
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rich-click==1.8.3
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
SCons==4.7.0
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.6.0
setproctitle==1.3.3
shapely==2.0.4
simpervisor==1.0.0
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
snakemake==8.14.0
snakemake-interface-common==1.17.2
snakemake-interface-executor-plugins==9.1.1
snakemake-interface-report-plugins==1.0.0
snakemake-interface-storage-plugins==3.2.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
spectra==0.0.11
Sphinx==7.4.3
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.31
stack-data==0.6.3
statsmodels==0.14.2
stopit==1.1.2
submitit==1.5.1
sympy==1.12.1
tables==3.9.2
tabulate==0.9.0
tenacity==8.4.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.0
tensorrt==10.1.0
tensorrt-cu12==10.1.0
tensorrt-cu12-bindings==10.1.0
tensorrt-cu12-libs==10.1.0
termcolor==2.4.0
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.5.0
throttler==1.2.2
tifffile==2024.6.18
tinycss2==1.3.0
tokenizers==0.19.1
toposort==1.10
torch==2.3.1
torch-fidelity==0.3.0
torch_cluster==1.6.3+pt22cu121
torch_geometric==2.5.3
torch_scatter==2.1.2+pt22cu121
torch_sparse==0.6.18+pt22cu121
torch_spline_conv==1.2.2+pt22cu121
torchaudio==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
traits==6.4.3
trame==3.6.3
trame-client==3.2.0
trame-jupyter-extension==2.1.2
trame-server==3.0.2
transformers==4.42.4
triton==2.3.1
typeguard==4.3.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
wandb==0.17.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.14.1
wslink==2.1.1
xgboost==2.1.0
y-py==0.6.2
yarl==1.9.4
ypy-websocket==0.8.4
yte==1.5.4
zipp==3.19.2
zope.interface==6.4.post2
